Fix Windows compile error.
BUG=
R=dslomov@chromium.org

Review URL: https://codereview.chromium.org/384743002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@22322 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/mark-compact.cc
@@ -4068,14 +4068,14 @@ intptr_t MarkCompactCollector::SweepConservatively(PagedSpace* space,
 
 
 int MarkCompactCollector::SweepInParallel(PagedSpace* space,
-                                          int required_freed_bytes) {
+                                          intptr_t required_freed_bytes) {
   PageIterator it(space);
   FreeList* free_list = space == heap()->old_pointer_space()
                             ? free_list_old_pointer_space_.get()
                             : free_list_old_data_space_.get();
   FreeList private_free_list(space);
-  int max_freed = 0;
-  int max_freed_overall = 0;
+  intptr_t max_freed = 0;
+  intptr_t max_freed_overall = 0;
   while (it.has_next()) {
     Page* p = it.next();
